app/compute/requirements.txt
Flask==0.10.1
jsonify==0.5
biopython==1.73
boto3==1.9.174